+++ slack-plugin/src/main/kotlin/slack/gradle/SlackProperties.kt
@@ -26,6 +26,7 @@ import slack.gradle.util.intProvider
 import slack.gradle.util.optionalStringProperty
 import slack.gradle.util.optionalStringProvider
 import slack.gradle.util.safeProperty
+import slack.gradle.util.sneakyNull
 
 /**
  * (Mostly Gradle) properties for configuration of SlackPlugin.
@@ -138,6 +139,14 @@ public class SlackProperties private constructor(private val project: Project) {
    */
   public val composeEnableLiveLiterals: Boolean
     get() = booleanProperty("slack.compose.android.enableLiveLiterals", false)
+
+  /**
+   * If true, uses the AndroidX compose compiler [SlackVersions.composeCompiler] for Compose
+   * Multiplatform compilations rather than the Jetbrains one. This can be useful in testing where
+   * AndroidX's compiler is farther ahead.
+   */
+  public val forceAndroidXComposeCompilerForComposeMultiplatform: Boolean
+    get() = booleanProperty("sgp.compose.multiplatform.forceAndroidXComposeCompiler", false)
 
   /**
    * When this property is present, the "internalRelease" build variant will have an application id
@@ -528,7 +537,18 @@ public class SlackProperties private constructor(private val project: Project) {
 
   /** Defines a required vendor for JDK toolchains. */
   public val jvmVendor: Provider<String>
-    get() = project.optionalStringProvider("sgp.config.jvmVendor")
+    get() =
+      project.optionalStringProvider("sgp.config.jvmVendor").map {
+        if (jvmVendorOptOut) {
+          sneakyNull()
+        } else {
+          it
+        }
+      }
+
+  /** Flag to disable JVM vendor setting locally. */
+  public val jvmVendorOptOut: Boolean
+    get() = booleanProperty("sgp.config.jvmVendor.optOut", defaultValue = false)
 
   internal fun requireAndroidSdkProperties(): AndroidSdkProperties {
     val compileSdk = compileSdkVersion ?: error("slack.compileSdkVersion not set")