+++ slack-plugin/src/main/kotlin/slack/gradle/StandardProjectConfigurations.kt
@@ -120,7 +120,7 @@ internal class StandardProjectConfigurations(
     val jdkVersion = project.jdkVersion()
     val jvmTargetVersion = project.jvmTargetVersion()
     project.applyJvmConfigurations(jdkVersion, jvmTargetVersion, slackProperties, slackExtension)
-    project.configureKotlinProjects(jdkVersion, jvmTargetVersion, slackProperties, slackExtension)
+    project.configureKotlinProjects(jdkVersion, jvmTargetVersion, slackProperties)
   }
 
   private fun Project.applyCommonConfigurations() {
@@ -828,7 +828,6 @@ internal class StandardProjectConfigurations(
     jdkVersion: Int?,
     jvmTargetVersion: Int,
     slackProperties: SlackProperties,
-    slackExtension: SlackExtension
   ) {
     val actualJvmTarget =
       if (jvmTargetVersion == 8) {
@@ -888,29 +887,6 @@ internal class StandardProjectConfigurations(
           }
           if (!isKaptGenerateStubsTask) {
             freeCompilerArgs.addAll(kotlinCompilerArgs)
-          }
-
-          if (slackExtension.featuresHandler.composeHandler.enabled.get()) {
-            logger.debug(
-              "Configuring compose compiler args in ${project.path}:${this@configureKotlinCompilationTask.name}"
-            )
-            if (!isKaptGenerateStubsTask) {
-              freeCompilerArgs.add("-Xskip-prerelease-check")
-            }
-            // Flag to disable Compose's kotlin version check because they're often behind
-            // Or ahead
-            // Or if they're the same, do nothing
-            // It's basically just very noisy.
-            val composeCompilerKotlinVersion =
-              slackProperties.versions.composeCompilerKotlinVersion
-                ?: error("Missing 'composeCompilerKotlinVersion' version in version catalog")
-            val kotlinVersion = slackProperties.versions.kotlin
-            if (!isKaptGenerateStubsTask && kotlinVersion != composeCompilerKotlinVersion) {
-              freeCompilerArgs.addAll(
-                "-P",
-                "plugin:androidx.compose.compiler.plugins.kotlin:suppressKotlinVersionCompatibilityCheck=$kotlinVersion"
-              )
-            }
           }
 
           if (this is KotlinJvmCompilerOptions) {